mailchimp widget title fix, uses apply_filters() instead of _e()
content in widgets appear to require apply_filters('widget_text') instead of _e() to translate their contents according to https://wpml.org/documentation/support/troubleshooting-string-localization/

The mailchimp plugin now appears to render the title as expected, the same should follow for other fields.

NOTE: the widget_text filter may be problematic when used with widgets that are not based on the default wordpress ones.